core/requirements.txt
Flask
docker
black
flask-sqlalchemy
python-dotenv 
flask-migrate 
mysqlclient
gunicorn
flask-login
flask-wtf
celery[redis]

#marshmallow-sqlalchemy